CD-6676 refactored code
